148: Move the Formatter wrapper classes into their own package
