+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/etor/demographics/PatientDemographicsController.java
@@ -3,10 +3,10 @@ package gov.hhs.cdc.trustedintermediary.etor.demographics;
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainRequest;
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainResponse;
 import gov.hhs.cdc.trustedintermediary.external.hapi.HapiDemographics;
-import gov.hhs.cdc.trustedintermediary.wrappers.Formatter;
-import gov.hhs.cdc.trustedintermediary.wrappers.FormatterProcessingException;
 import gov.hhs.cdc.trustedintermediary.wrappers.HapiFhir;
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger;
+import gov.hhs.cdc.trustedintermediary.wrappers.formatter.Formatter;
+import gov.hhs.cdc.trustedintermediary.wrappers.formatter.FormatterProcessingException;
 import java.util.Map;
 import java.util.Optional;
 import javax.inject.Inject;
