--- app/src/main/java/gov/hhs/cdc/trustedintermediary/wrappers/TypeReference.java
@@ -1,3 +1,0 @@
-package gov.hhs.cdc.trustedintermediary.wrappers;
-
-public abstract class TypeReference<T> {}
