+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/wrappers/formatter/TypeReference.java
@@ -1,0 +1,3 @@
+package gov.hhs.cdc.trustedintermediary.wrappers.formatter;
+
+public abstract class TypeReference<T> {}
